Influence function estimates for SPVIMs

Compute the influence functions for the contribution from sampling observations and subsets.

spvim_ics(Z, z_counts, W, v, psi, G, c_n, ics, measure)

Arguments

Z

the matrix of presence/absence of each feature (columns) in each sampled subset (rows)

z_counts

the number of times each unique subset was sampled

W

the matrix of weights

v

the estimated predictiveness measures

psi

the estimated SPVIM values

G

the constraint matrix

c_n

the constraint values

ics

a list of influence function values for each predictiveness measure

measure

the type of measure (e.g., "r_squared" or "auc")

Value

a named list of length 2; contrib_v is the contribution from estimating V, while contrib_s is the contribution from sampling subsets.

Details

The processes for sampling observations and sampling subsets are independent. Thus, we can compute the influence function separately for each sampling process. For further details, see the paper by Williamson and Feng (2020).