vimp: Perform Inference on Algorithm-Agnostic Variable Importance

A unified framework for valid statistical inference on algorithm-agnostic measures of variable importance. You provide the data, a method for estimating the conditional mean of the outcome given the covariates, choose a variable importance measure, and specify variable(s) of interest; 'vimp' takes care of the rest.

Author(s)

Maintainer: Brian Williamson http://bdwilliamson.github.io

Methodology authors:

Brian D. Williamson

Peter B. Gilbert

Noah R. Simon

Marco Carone

See Also

Preprints:

http://biostats.bepress.com/uwbiostat/paper422/ (R-squared-based variable importance)

http://arxiv.org/abs/2004.03683 (general variable importance)

https://arxiv.org/abs/2006.09481 (general Shapley-based variable importance)

Other useful links:

http://bdwilliamson.github.io/vimp

http://github.com/bdwilliamson/vimp

Report bugs at http://github.com/bdwilliamson/vimp/issues

Imports

The packages that we import either make the internal code nice (dplyr, magrittr, tibble, rlang, MASS), are directly relevant to estimating the conditional mean (SuperLearner) or predictiveness measures (ROCR), or are necessary for hypothesis testing (stats).

We suggest several other packages: xgboost, ranger, gam, glmnet, polspline, and quadprog allow a flexible library of candidate learners in the Super Learner; ggplot2, cowplot, and forcats help with plotting variable importance estimates; testthat and covr help with unit tests; and knitr, rmarkdown, and RCurl help with the vignettes and examples.