Confidence intervals for variable importance

Compute confidence intervals for the true variable importance parameter.

vimp_ci(est, se, scale = "identity", level = 0.95, truncate = TRUE)

Arguments

est

estimate of variable importance, e.g., from a call to vimp_point_est.

se

estimate of the standard error of est, e.g., from a call to vimp_se.

scale

scale to compute interval estimate on (defaults to "identity": compute Wald-type CI).

level

confidence interval type (defaults to 0.95).

truncate

truncate CIs to have lower limit at (or above) zero?

Value

The Wald-based confidence interval for the true importance of the given group of left-out covariates.

Details

See the paper by Williamson, Gilbert, Simon, and Carone for more details on the mathematics behind this function and the definition of the parameter of interest.