Check pre-computed fitted values for call to vim, cv_vim, or sp_vim

Check pre-computed fitted values for call to vim, cv_vim, or sp_vim

check_fitted_values(
  Y = NULL,
  f1 = NULL,
  f2 = NULL,
  cross_fitted_f1 = NULL,
  cross_fitted_f2 = NULL,
  sample_splitting_folds = NULL,
  cross_fitting_folds = NULL,
  V = NULL,
  ss_V = NULL,
  cv = FALSE
)

Arguments

Y
the outcome

f1
estimator of the population-optimal prediction function using all covariates

f2
estimator of the population-optimal prediction function using the reduced set of covariates

cross_fitted_f1
cross-fitted estimator of the population-optimal prediction function using all covariates

cross_fitted_f2
cross-fitted estimator of the population-optimal prediction function using the reduced set of covariates

sample_splitting_folds
the folds for sample-splitting (used for hypothesis testing)

cross_fitting_folds
the folds for cross-fitting (used for point estimates of variable importance in cv_vim and sp_vim)

V
the number of cross-fitting folds

ss_V
the number of folds for CV (if sample_splitting is TRUE)

cv
a logical flag indicating whether or not to use cross-fitting

X
the covariates

Value

None. Called for the side effect of stopping the algorithm if any inputs are in an unexpected format.

Details

Ensure that inputs to vim, cv_vim, and sp_vim follow the correct formats.