Shapley Population Variable Importance Measure (SPVIM) Estimates and Inference

Compute estimates and confidence intervals for the SPVIMs, using cross-fitting.

sp_vim(
  Y = NULL,
  X = NULL,
  V = 5,
  type = "r_squared",
  SL.library = c("SL.glmnet", "SL.xgboost", "SL.mean"),
  univariate_SL.library = NULL,
  gamma = 1,
  alpha = 0.05,
  delta = 0,
  na.rm = FALSE,
  stratified = FALSE,
  verbose = FALSE,
  C = rep(1, length(Y)),
  Z = NULL,
  ipc_weights = rep(1, length(Y)),
  ...
)

Arguments

Y
the outcome.

X
the covariates.

V
the number of folds for cross-fitting, defaults to 10.

type
the type of parameter (e.g., R-squared-based is "r_squared"). Note that type = 'anova' is not allowed for SPVIMs.

SL.library
a character vector of learners to pass to SuperLearner, if f1 and f2 are Y and X, respectively. Defaults to SL.glmnet, SL.xgboost, and SL.mean.

univariate_SL.library
(optional) a character vector of learners to pass to SuperLearner for estimating univariate regression functions. Defaults to SL.polymars

gamma
the fraction of the sample size to use when sampling subsets (e.g., gamma = 1 samples the same number of subsets as the sample size)

alpha
the level to compute the confidence interval at. Defaults to 0.05, corresponding to a 95% confidence interval.

delta
the value of the \(\delta\)-null (i.e., testing if importance < \(\delta\)); defaults to 0.

na.rm
should we remove NA's in the outcome and fitted values in computation? (defaults to FALSE)

stratified
should the generated folds be stratified based on the outcome (helps to ensure class balance across cross-fitting folds)?

verbose
should sp_vim and SuperLearner print out progress? (defaults to FALSE)

C
the indicator of coarsening (1 denotes observed, 0 denotes unobserved).

Z
either (i) NULL (the default, in which case the argument C above must be all ones), or (ii) a character vector specifying the variable(s) among Y and X that are thought to play a role in the coarsening mechanism.

ipc_weights
weights for the computed influence curve (i.e., inverse probability weights for coarsened-at-random settings). Assumed to be already inverted (i.e., ipc_weights = 1 / [estimated probability weights]).

...
other arguments to the estimation tool, see "See also".

Value

An object of class vim. See Details for more information.

Details

We define the SPVIM as the weighted average of the population difference in predictiveness over all subsets of features not containing feature \(j\).

This is equivalent to finding the solution to a population weighted least squares problem. This key fact allows us to estimate the SPVIM using weighted least squares, where we first sample subsets from the power set of all possible features using the Shapley sampling distribution; then use cross-fitting to obtain estimators of the predictiveness of each sampled subset; and finally, solve the least squares problem given in Williamson and Feng (2020).

See the paper by Williamson and Feng (2020) for more details on the mathematics behind this function, and the validity of the confidence intervals. The function works by estimating In the interest of transparency, we return most of the calculations within the vim object. This results in a list containing:

SL.library - the library of learners passed to SuperLearner

v- the estimated predictiveness measure for each sampled subset

preds_lst - the predicted values from the chosen method for each sampled subset

est - the estimated SPVIM value for each feature

ic_lst - the influence functions for each sampled subset

ic- a list of the SPVIM influence function contributions

se - the standard errors for the estimated variable importance

ci - the \((1-\alpha) \times 100\)% confidence intervals based on the variable importance estimates

gamma- the fraction of the sample size used when sampling subsets

alpha - the level, for confidence interval calculation

delta- the delta value used for hypothesis testing

y - the outcome

ipc_weights - the weights

mat- a tibble with the estimates, SEs, CIs, hypothesis testing decisions, and p-values

See also

Examples

library(SuperLearner)
library(ranger)
n <- 100
p <- 2
# generate the data
x <- data.frame(replicate(p, stats::runif(n, -5, 5)))

# apply the function to the x's
smooth <- (x[,1]/5)^2*(x[,1]+7)/5 + (x[,2]/3)^2

# generate Y ~ Normal (smooth, 1)
y <- as.matrix(smooth + stats::rnorm(n, 0, 1))

# set up a library for SuperLearner
learners <- c("SL.mean", "SL.ranger")

# -----------------------------------------
# using Super Learner (with a small number of CV folds,
# for illustration only)
# -----------------------------------------
set.seed(4747)
est <- sp_vim(Y = y, X = x, V = 2, type = "r_squared",
SL.library = learners, alpha = 0.05)

#> Warning: number of items to replace is not a multiple of replacement length

#> Warning: number of items to replace is not a multiple of replacement length

#> Warning: number of items to replace is not a multiple of replacement length

#> Warning: One or more original estimates < 0; returning zero for these indices.

#> Error in tcrossprod(x, y): requires numeric/complex matrix/vector arguments