Nonparametric Variable Importance Estimates and Inference

Compute estimates of and confidence intervals for nonparametric risk-based variable importance.

vim(
  Y,
  X,
  f1 = NULL,
  f2 = NULL,
  indx = 1,
  weights = rep(1, length(Y)),
  type = "r_squared",
  run_regression = TRUE,
  SL.library = c("SL.glmnet", "SL.xgboost", "SL.mean"),
  alpha = 0.05,
  delta = 0,
  scale = "identity",
  na.rm = FALSE,
  folds = NULL,
  stratified = FALSE,
  ...
)

Arguments

Y
the outcome.

X
the covariates.

f1
the fitted values from a flexible estimation technique regressing Y on X.

f2
the fitted values from a flexible estimation technique regressing Y on X withholding the columns in indx.

indx
the indices of the covariate(s) to calculate variable importance for; defaults to 1.

weights
weights for the computed influence curve (e.g., inverse probability weights for coarsened-at-random settings)

type
the type of importance to compute; defaults to r_squared, but other supported options are auc, accuracy, and anova.

run_regression
if outcome Y and covariates X are passed to vimp_accuracy, and run_regression is TRUE, then Super Learner will be used; otherwise, variable importance will be computed using the inputted fitted values.

SL.library
a character vector of learners to pass to SuperLearner, if f1 and f2 are Y and X, respectively. Defaults to SL.glmnet, SL.xgboost, and SL.mean.

alpha
the level to compute the confidence interval at. Defaults to 0.05, corresponding to a 95% confidence interval.

delta
the value of the \(\delta\)-null (i.e., testing if importance < \(\delta\)); defaults to 0.

scale
should CIs be computed on original ("identity") or logit ("logit") scale?

na.rm
should we remove NA's in the outcome and fitted values in computation? (defaults to FALSE)

folds
the folds used for f1 and f2; assumed to be 1 for the observations used in f1 and 2 for the observations used in f2. If there is only a single fold passed in, then hypothesis testing is not done.

stratified
if run_regression = TRUE, then should the generated folds be stratified based on the outcome (helps to ensure class balance across cross-validation folds)

...
other arguments to the estimation tool, see "See also".

Value

An object of classes vim and the type of risk-based measure. See Details for more information.

Details

We define the population variable importance measure (VIM) for the group of features (or single feature) \(s\) with respect to the predictiveness measure \(V\) by $$\psi_{0,s} := V(f_0, P_0) - V(f_{0,s}, P_0),$$ where \(f_0\) is the population predictiveness maximizing function, \(f_{0,s}\) is the population predictiveness maximizing function that is only allowed to access the features with index not in \(s\), and \(P_0\) is the true data-generating distribution. VIM estimates are obtained by obtaining estimators \(f_n\) and \(f_{n,s}\) of \(f_0\) and \(f_{0,s}\), respectively; obtaining an estimator \(P_n\) of \(P_0\); and finally, setting \(\psi_{n,s} := V(f_n, P_n) - V(f_{n,s}, P_n)\).

In the interest of transparency, we return most of the calculations within the vim object. This results in a list containing:

call - the call to vim

s - the column(s) to calculate variable importance for

SL.library - the library of learners passed to SuperLearner

type - the type of risk-based variable importance measured

full_fit - the fitted values of the chosen method fit to the full data

red_fit - the fitted values of the chosen method fit to the reduced data

est - the estimated variable importance

naive - the naive estimator of variable importance

update - the influence curve-based update

se - the standard error for the estimated variable importance

ci - the \((1-\alpha) \times 100\)% confidence interval for the variable importance estimate

test - a decision to either reject (TRUE) or not reject (FALSE) the null hypothesis, based on a conservative test

pval - a conservative p-value based on the same conservative test as test

full_mod - the object returned by the estimation procedure for the full data regression (if applicable)

red_mod - the object returned by the estimation procedure for the reduced data regression (if applicable)

alpha - the level, for confidence interval calculation

folds - the folds used for hypothesis testing

y - the outcome

weights - the weights

mat- a tibble with the estimate, SE, CI, hypothesis testing decision, and p-value

See also

Examples

library(SuperLearner)
library(ranger)
## generate the data
## generate X
p <- 2
n <- 100
x <- data.frame(replicate(p, stats::runif(n, -1, 1)))

## apply the function to the x's
f <- function(x) 0.5 + 0.3*x[1] + 0.2*x[2]
smooth <- apply(x, 1, function(z) f(z))

## generate Y ~ Normal (smooth, 1)
y <- matrix(rbinom(n, size = 1, prob = smooth))

## set up a library for SuperLearner
learners <- "SL.ranger"

## using Y and X; use class-balanced folds
folds_1 <- sample(rep(seq_len(2), length = sum(y == 1)))
folds_0 <- sample(rep(seq_len(2), length = sum(y == 0)))
folds <- vector("numeric", length(y))
folds[y == 1] <- folds_1
folds[y == 0] <- folds_0
est <- vim(y, x, indx = 2, type = "r_squared",
           alpha = 0.05, run_regression = TRUE,
           SL.library = learners, cvControl = list(V = 2),
           folds = folds)

#> Error in data.frame(..., check.names = FALSE) : 
#>   arguments imply differing number of rows: 26, 24

#> Warning: Error in algorithm SL.ranger 
#>   The Algorithm will be removed from the Super Learner (i.e. given weight 0) 

#> Error in data.frame(..., check.names = FALSE) : 
#>   arguments imply differing number of rows: 27, 25

#> Warning: Error in algorithm SL.ranger 
#>   The Algorithm will be removed from the Super Learner (i.e. given weight 0) 

#> Error in (function (Y, X, newX = NULL, family = gaussian(), SL.library,     method = "method.NNLS", id = NULL, verbose = FALSE, control = list(),     cvControl = list(), obsWeights = NULL, env = parent.frame()) {    time_start = proc.time()    if (is.character(method)) {        if (exists(method, mode = "list")) {            method <- get(method, mode = "list")        }        else if (exists(method, mode = "function")) {            method <- get(method, mode = "function")()        }    }    else if (is.function(method)) {        method <- method()    }    if (!is.list(method)) {        stop("method is not in the appropriate format. Check out help('method.template')")    }    if (!is.null(method$require)) {        sapply(method$require, function(x) require(force(x),             character.only = TRUE))    }    control <- do.call("SuperLearner.control", control)    cvControl <- do.call("SuperLearner.CV.control", cvControl)    library <- .createLibrary(SL.library)    .check.SL.library(library = c(unique(library$library$predAlgorithm),         library$screenAlgorithm))    call <- match.call(expand.dots = TRUE)    if (!inherits(X, "data.frame"))         message("X is not a data frame. Check the algorithms in SL.library to make sure they are compatible with non data.frame inputs")    varNames <- colnames(X)    N <- dim(X)[1L]    p <- dim(X)[2L]    k <- nrow(library$library)    kScreen <- length(library$screenAlgorithm)    Z <- matrix(NA, N, k)    libraryNames <- paste(library$library$predAlgorithm, library$screenAlgorithm[library$library$rowScreen],         sep = "_")    if (p < 2 & !identical(library$screenAlgorithm, "All")) {        warning("Screening algorithms specified in combination with single-column X.")    }    fitLibEnv <- new.env()    assign("fitLibrary", vector("list", length = k), envir = fitLibEnv)    assign("libraryNames", libraryNames, envir = fitLibEnv)    evalq(names(fitLibrary) <- libraryNames, envir = fitLibEnv)    errorsInCVLibrary <- rep(0, k)    errorsInLibrary <- rep(0, k)    if (is.null(newX)) {        newX <- X    }    if (!identical(colnames(X), colnames(newX))) {        stop("The variable names and order in newX must be identical to the variable names and order in X")    }    if (sum(is.na(X)) > 0 | sum(is.na(newX)) > 0 | sum(is.na(Y)) >         0) {        stop("missing data is currently not supported. Check Y, X, and newX for missing values")    }    if (!is.numeric(Y)) {        stop("the outcome Y must be a numeric vector")    }    if (is.character(family))         family <- get(family, mode = "function", envir = parent.frame())    if (is.function(family))         family <- family()    if (is.null(family$family)) {        print(family)        stop("'family' not recognized")    }    if (family$family != "binomial" & isTRUE("cvAUC" %in% method$require)) {        stop("'method.AUC' is designed for the 'binomial' family only")    }    validRows <- CVFolds(N = N, id = id, Y = Y, cvControl = cvControl)    if (is.null(id)) {        id <- seq(N)    }    if (!identical(length(id), N)) {        stop("id vector must have the same dimension as Y")    }    if (is.null(obsWeights)) {        obsWeights <- rep(1, N)    }    if (!identical(length(obsWeights), N)) {        stop("obsWeights vector must have the same dimension as Y")    }    .crossValFUN <- function(valid, Y, dataX, id, obsWeights,         library, kScreen, k, p, libraryNames, saveCVFitLibrary) {        tempLearn <- dataX[-valid, , drop = FALSE]        tempOutcome <- Y[-valid]        tempValid <- dataX[valid, , drop = FALSE]        tempWhichScreen <- matrix(NA, nrow = kScreen, ncol = p)        tempId <- id[-valid]        tempObsWeights <- obsWeights[-valid]        for (s in seq(kScreen)) {            screen_fn = get(library$screenAlgorithm[s], envir = env)            testScreen <- try(do.call(screen_fn, list(Y = tempOutcome,                 X = tempLearn, family = family, id = tempId,                 obsWeights = tempObsWeights)))            if (inherits(testScreen, "try-error")) {                warning(paste("replacing failed screening algorithm,",                   library$screenAlgorithm[s], ", with All()",                   "\n "))                tempWhichScreen[s, ] <- TRUE            }            else {                tempWhichScreen[s, ] <- testScreen            }            if (verbose) {                message(paste("Number of covariates in ", library$screenAlgorithm[s],                   " is: ", sum(tempWhichScreen[s, ]), sep = ""))            }        }        out <- matrix(NA, nrow = nrow(tempValid), ncol = k)        if (saveCVFitLibrary) {            model_out <- vector(mode = "list", length = k)        }        else {            model_out <- NULL        }        for (s in seq(k)) {            pred_fn = get(library$library$predAlgorithm[s], envir = env)            testAlg <- try(do.call(pred_fn, list(Y = tempOutcome,                 X = subset(tempLearn, select = tempWhichScreen[library$library$rowScreen[s],                   ], drop = FALSE), newX = subset(tempValid,                   select = tempWhichScreen[library$library$rowScreen[s],                     ], drop = FALSE), family = family, id = tempId,                 obsWeights = tempObsWeights)))            if (inherits(testAlg, "try-error")) {                warning(paste("Error in algorithm", library$library$predAlgorithm[s],                   "\n  The Algorithm will be removed from the Super Learner (i.e. given weight 0) \n"))            }            else {                out[, s] <- testAlg$pred                if (saveCVFitLibrary) {                  model_out[[s]] <- testAlg$fit                }            }            if (verbose)                 message(paste("CV", libraryNames[s]))        }        if (saveCVFitLibrary) {            names(model_out) <- libraryNames        }        invisible(list(out = out, model_out = model_out))    }    time_train_start = proc.time()    crossValFUN_out <- lapply(validRows, FUN = .crossValFUN,         Y = Y, dataX = X, id = id, obsWeights = obsWeights, library = library,         kScreen = kScreen, k = k, p = p, libraryNames = libraryNames,         saveCVFitLibrary = control$saveCVFitLibrary)    Z[unlist(validRows, use.names = FALSE), ] <- do.call("rbind",         lapply(crossValFUN_out, "[[", "out"))    if (control$saveCVFitLibrary) {        cvFitLibrary <- lapply(crossValFUN_out, "[[", "model_out")    }    else {        cvFitLibrary <- NULL    }    errorsInCVLibrary <- apply(Z, 2, function(x) anyNA(x))    if (sum(errorsInCVLibrary) > 0) {        Z[, as.logical(errorsInCVLibrary)] <- 0    }    if (all(Z == 0)) {        stop("All algorithms dropped from library")    }    getCoef <- method$computeCoef(Z = Z, Y = Y, libraryNames = libraryNames,         obsWeights = obsWeights, control = control, verbose = verbose,         errorsInLibrary = errorsInCVLibrary)    coef <- getCoef$coef    names(coef) <- libraryNames    time_train = proc.time() - time_train_start    if (!("optimizer" %in% names(getCoef))) {        getCoef["optimizer"] <- NA    }    m <- dim(newX)[1L]    predY <- matrix(NA, nrow = m, ncol = k)    .screenFun <- function(fun, list) {        screen_fn = get(fun, envir = env)        testScreen <- try(do.call(screen_fn, list))        if (inherits(testScreen, "try-error")) {            warning(paste("replacing failed screening algorithm,",                 fun, ", with All() in full data", "\n "))            out <- rep(TRUE, ncol(list$X))        }        else {            out <- testScreen        }        return(out)    }    time_predict_start = proc.time()    whichScreen <- sapply(library$screenAlgorithm, FUN = .screenFun,         list = list(Y = Y, X = X, family = family, id = id, obsWeights = obsWeights),         simplify = FALSE)    whichScreen <- do.call(rbind, whichScreen)    .predFun <- function(index, lib, Y, dataX, newX, whichScreen,         family, id, obsWeights, verbose, control, libraryNames) {        pred_fn = get(lib$predAlgorithm[index], envir = env)        testAlg <- try(do.call(pred_fn, list(Y = Y, X = subset(dataX,             select = whichScreen[lib$rowScreen[index], ], drop = FALSE),             newX = subset(newX, select = whichScreen[lib$rowScreen[index],                 ], drop = FALSE), family = family, id = id, obsWeights = obsWeights)))        if (inherits(testAlg, "try-error")) {            warning(paste("Error in algorithm", lib$predAlgorithm[index],                 " on full data", "\n  The Algorithm will be removed from the Super Learner (i.e. given weight 0) \n"))            out <- rep.int(NA, times = nrow(newX))        }        else {            out <- testAlg$pred            if (control$saveFitLibrary) {                eval(bquote(fitLibrary[[.(index)]] <- .(testAlg$fit)),                   envir = fitLibEnv)            }        }        if (verbose) {            message(paste("full", libraryNames[index]))        }        invisible(out)    }    predY <- do.call("cbind", lapply(seq(k), FUN = .predFun,         lib = library$library, Y = Y, dataX = X, newX = newX,         whichScreen = whichScreen, family = family, id = id,         obsWeights = obsWeights, verbose = verbose, control = control,         libraryNames = libraryNames))    errorsInLibrary <- apply(predY, 2, function(algorithm) anyNA(algorithm))    if (sum(errorsInLibrary) > 0) {        if (sum(coef[as.logical(errorsInLibrary)]) > 0) {            warning(paste0("Re-running estimation of coefficients removing failed algorithm(s)\n",                 "Original coefficients are: \n", paste(coef,                   collapse = ", "), "\n"))            Z[, as.logical(errorsInLibrary)] <- 0            if (all(Z == 0)) {                stop("All algorithms dropped from library")            }            getCoef <- method$computeCoef(Z = Z, Y = Y, libraryNames = libraryNames,                 obsWeights = obsWeights, control = control, verbose = verbose,                 errorsInLibrary = errorsInLibrary)            coef <- getCoef$coef            names(coef) <- libraryNames        }        else {            warning("Coefficients already 0 for all failed algorithm(s)")        }    }    getPred <- method$computePred(predY = predY, coef = coef,         control = control)    time_predict = proc.time() - time_predict_start    colnames(predY) <- libraryNames    if (sum(errorsInCVLibrary) > 0) {        getCoef$cvRisk[as.logical(errorsInCVLibrary)] <- NA    }    time_end = proc.time()    times = list(everything = time_end - time_start, train = time_train,         predict = time_predict)    out <- list(call = call, libraryNames = libraryNames, SL.library = library,         SL.predict = getPred, coef = coef, library.predict = predY,         Z = Z, cvRisk = getCoef$cvRisk, family = family, fitLibrary = get("fitLibrary",             envir = fitLibEnv), cvFitLibrary = cvFitLibrary,         varNames = varNames, validRows = validRows, method = method,         whichScreen = whichScreen, control = control, cvControl = cvControl,         errorsInCVLibrary = errorsInCVLibrary, errorsInLibrary = errorsInLibrary,         metaOptimizer = getCoef$optimizer, env = env, times = times)    class(out) <- c("SuperLearner")    return(out)})(cvControl = list(V = 2), family = structure(list(family = "gaussian",     link = "identity", linkfun = function (mu)     mu, linkinv = function (eta)     eta, variance = function (mu)     rep.int(1, length(mu)), dev.resids = function (y, mu, wt)     wt * ((y - mu)^2), aic = function (y, n, mu, wt, dev)     {        nobs <- length(y)        nobs * (log(dev/nobs * 2 * pi) + 1) + 2 - sum(log(wt))    }, mu.eta = function (eta)     rep.int(1, length(eta)), initialize = expression({        n <- rep.int(1, nobs)        if (is.null(etastart) && is.null(start) && is.null(mustart) &&             ((family$link == "inverse" && any(y == 0)) || (family$link ==                 "log" && any(y <= 0))))             stop("cannot find valid starting values: please specify some")        mustart <- y    }), validmu = function (mu)     TRUE, valideta = function (eta)     TRUE), class = "family"), Y = structure(c(0.716033333333334, 0.8915, 0.791433333333334, 0.3265, 0.233966666666667, 0.824800000000001, 0.630566666666666, 0.0326333333333333, 0.8264, 0.729266666666667, 0.274666666666667, 0.782066666666667, 0.474233333333333, 0.2663, 0.713166666666667, 0.303466666666667, 0.6997, 0.312866666666667, 0.4274, 0.511366666666667, 0.6662, 0.5423, 0.573966666666666, 0.347166666666667, 0.259433333333333, 0.714833333333333, 0.239533333333333, 0.677066666666667, 0.210166666666667, 0.0943333333333334, 0.696333333333333, 0.826433333333333, 0.635, 0.219133333333333, 0.594533333333333, 0.234366666666667, 0.0876333333333333, 0.280566666666667, 0.3992, 0.1142, 0.2058, 0.716066666666667, 0.0433333333333333, 0.0415666666666666, 0.0445666666666667, 0.0836, 0.8141, 0.0663666666666667, 0.7699, 0.6073, 0.8933), .Dim = c(51L, 1L)), X = structure(list(X1 = c(-0.753771381918341, -0.143795292358845, 0.285397629719228, -0.598208559677005, 0.152677124366164, 0.868483573198318, -0.59913653600961, -0.386898460797966, 0.669352719094604, 0.894915781915188, -0.539103284012526, 0.894362262450159, -0.305378235410899, -0.757027229294181, -0.376907113939524, 0.168692866805941, -0.455501619726419, -0.358561683446169, -0.114100332371891, 0.814544485881925, 0.482119514606893, 0.620705139823258, 0.319245150778443, -0.0643384926952422, 0.865729700773954, -0.0245725363492966, 0.479926395229995, -0.598868171684444, 0.0541091421619058, 0.705140637699515, -0.679088393226266, -0.110585407819599, -0.594418883323669, 0.954369420185685, -0.768858699128032, -0.612607419025153, 0.19230820890516, 0.623889425769448, 0.208979279734194, -0.532531827688217, 0.240859299432486, -0.14225233765319, -0.946452054195106, -0.518148417584598, -0.507158214226365, -0.683654933236539, 0.962758276611567, -0.480572562664747, -0.846726555377245)), row.names = c(2L, 3L, 5L, 6L, 13L, 20L, 21L, 22L, 24L, 28L, 29L, 30L, 32L, 35L, 36L, 38L, 39L, 40L, 43L, 46L, 47L, 48L, 51L, 53L, 56L, 58L, 66L, 69L, 72L, 73L, 74L, 76L, 77L, 78L, 79L, 80L, 81L, 82L, 83L, 85L, 86L, 87L, 88L, 89L, 90L, 92L, 96L, 97L, 100L), class = "data.frame"), SL.library = "SL.ranger"): All algorithms dropped from library

## using pre-computed fitted values
full <- SuperLearner(Y = y[folds == 1], X = x[folds == 1, ],
SL.library = learners, cvControl = list(V = 2))
full.fit <- predict(full)$pred
reduced <- SuperLearner(Y = y[folds == 2], X = x[folds == 2, -2, drop = FALSE],
SL.library = learners, cvControl = list(V = 2))
red.fit <- predict(reduced)$pred

est <- vim(Y = y, f1 = full.fit, f2 = red.fit,
            indx = 2, run_regression = FALSE, alpha = 0.05, folds = folds,
            type = "accuracy")